Fix makefiles

+++ assembler/src/utilities.c
@@ -8,7 +8,7 @@ void strzero(char *s) {
 }
 
 void leftShift(char *s) {
-    for(int i = 1; i < strlen(s); i++) {
+    for(int i = 1; i < (int)strlen(s); i++) {
         s[i-1] = s[i];
     }
     s[strlen(s)-1]='\0';
@@ -16,8 +16,8 @@ void leftShift(char *s) {
 
 int strcontains(char *s, const char *c) {
     int exit=0;
-    for(int i=0; i<strlen(s) && exit!=1; ++i) {
-        for(int x=0; x<strlen(c); ++x) {
+    for(int i=0; i<(int)strlen(s) && exit!=1; ++i) {
+        for(int x=0; x<(int)strlen(c); ++x) {
             if(s[i]==c[x]) {
                 exit=1;
             }
@@ -29,7 +29,7 @@ int strcontains(char *s, const char *c) {
 
 int str2int(char *s) {
     int value=0;
-    for(int i=0; i<strlen(s); ++i) {
+    for(int i=0; i<(int)strlen(s); ++i) {
         value = value*10+(s[i]-'0');
     }
     return value;
